+++ src/models/ratingSystem.ts
@@ -1,40 +1,51 @@
-import {Sdk, SdkConfig} from "../config/sdk";
-import {Address} from "viem";
-import abi from "../../abis/RatingSystem.json"
+import Addresses from "../../contractAddresses.json";
+import { Sdk, SdkConfig } from "../config/sdk";
+import abi from "../../abis/RatingSystem.json";
 
-const RatingSystemABI: typeof abi = abi
+const ContractAddresses: typeof Addresses = Addresses;
+const RatingSystemABI: typeof abi = abi;
 
 export class RatingSystem {
-    private config: SdkConfig
-    constructor(sdkInstance: Sdk) {
-        this.config = sdkInstance.config
-    }
+  private config: SdkConfig;
+  constructor(sdkInstance: Sdk) {
+    this.config = sdkInstance.config;
 
-    async view(address: Address, functionName: string, args: any) {
-        try {
-            return await this.config.publicClient.readContract({
-                address: address,
-                abi: RatingSystemABI.abi,
-                functionName: functionName,
-                args: args
-            })
-        } catch (error) {
-            throw error
-        }
+    if (
+      ContractAddresses.RatingSystem[this.config.publicClient.chain!.id] ===
+      undefined
+    ) {
+      throw new Error(
+        "RatingSystem address not found for network id: " +
+          this.config.publicClient.chain!.id
+      );
     }
+  }
 
-    async execute(address: Address, functionName: string, args: any) {
-        try {
-            const { request } = await this.config.publicClient.simulateContract({
-                address: address,
-                abi: RatingSystemABI.abi,
-                functionName: functionName,
-                args: args,
-                account: this.config.walletClient.account!.address,
-            })
-            return await this.config.walletClient.writeContract(request)
-        } catch (error) {
-            throw error
-        }
+  async view(functionName: string, args: any[]): Promise<any> {
+    try {
+      return await this.config.publicClient.readContract({
+        address: ContractAddresses.RatingSystem[this.config.publicClient.chain!.id],
+        abi: RatingSystemABI.abi,
+        functionName: functionName,
+        args: args,
+      });
+    } catch (error) {
+      throw error;
     }
+  }
+
+  async execute(functionName: string, args: any[]): Promise<any> {
+    try {
+      const { request } = await this.config.publicClient.simulateContract({
+        address: ContractAddresses.RatingSystem[this.config.publicClient.chain!.id],
+        abi: RatingSystemABI.abi,
+        functionName: functionName,
+        args: args,
+        account: this.config.walletClient.account,
+      });
+      return await this.config.walletClient.writeContract(request);
+    } catch (error) {
+      throw error;
+    }
+  }
 }